# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GetCarApi — vehicle data collection and VIN history API
servers:
  - url: /api
    description: Base API path
...
security: []   # No global security — applied per-endpoint below
paths:
...
  /v1/vin/check/{vin}:
    get:
      operationId: vinCheck
      tags: [vin]
      summary: Check VIN existence (Bearer required, no credit)
      description: >
        Returns whether the VIN exists in the database along with which
        data providers have records for it and whether historical data is
        available. Requires a valid API token. Does not consume a credit.
      security:
        - bearerAuth: []
      parameters:
        - name: vin
          in: path
          required: true
          schema:
            type: string
            minLength: 5
            maxLength: 17
          description: Vehicle Identification Number (case-insensitive, 5–17 chars)
          example: KNAGM4A72E5158989
      responses:
        "200":
          description: VIN check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VinCheckEnvelope"
              example:
                success: true
                data:
                  vin: KNAGM4A72E5158989
                  exists: true
                  providers: [encar]
                  hasHistory: true
        "400":
          description: Invalid VIN format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicErrorResponse"
        "401":
          description: Missing, invalid, or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicErrorResponse"
        "403":
          description: Token not permitted for this route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicErrorResponse"
        "503":
          description: VIN check temporarily disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicErrorResponse"
  /v1/vin/{vin}:
    get:
      operationId: vinHistory
      tags: [vin]
      summary: Full vehicle history (token-authenticated)
      description: >
        Returns the complete vehicle history for a VIN: core vehicle
        attributes, all listing records from every provider, price and
        mileage observations, lifecycle events, and available photos.


        **Authentication:** `Authorization: Bearer <token>`


        **Credits:** One credit ($2 USD) is consumed per successful (200) response.
        Requests that return 404 (VIN not found), 402 (no credits), or 429 (rate limited)
        do NOT consume a credit.


        Remaining quotas and credit balances are available in the client portal only —
        they are not returned on API responses.
      security:
        - bearerAuth: []
      parameters:
        - name: vin
          in: path
          required: true
          schema:
            type: string
            minLength: 5
            maxLength: 17
          description: Vehicle Identification Number (case-insensitive)
          example: KNAGM4A72E5158989
      responses:
        "200":
          description: Full vehicle history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VinHistoryEnvelope"
        "400":
          description: Invalid VIN format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicErrorResponse"
        "401":
          description: Missing, invalid, or expired API token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicErrorResponse"
              examples:
                missingToken:
                  value:
                    success: false
                    error:
                      code: MISSING_TOKEN
                      message: "API token required. Provide: Authorization: Bearer <token>"
                invalidToken:
                  value:
                    success: false
                    error:
                      code: INVALID_TOKEN
                      message: Invalid or revoked API token
                tokenExpired:
                  value:
                    success: false
                    error:
                      code: TOKEN_EXPIRED
                      message: API token has expired
        "404":
          description: VIN not found in the database
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicErrorResponse"
              example:
                success: false
                error:
                  code: VIN_NOT_FOUND
                  message: No vehicle found for this VIN
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicErrorResponse"
              examples:
                dailyLimit:
                  value:
                    success: false
                    error:
                      code: DAILY_LIMIT_EXCEEDED
                      message: Daily request limit exceeded
                vinLimit:
                  value:
                    success: false
                    error:
                      code: VIN_LIMIT_EXCEEDED
                      message: Per-VIN request limit exceeded for this billing period

  # ── System Events (Observability) ────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: VDI token (vdi_…)
      description: >
        API token issued to your account. Pass as
        `Authorization: Bearer vdi_<token>`. Required for all VIN routes
        (including check) and live inventory endpoints.
  schemas:
...
    AuctionSaleRow:
      type: object
      required: [soldDate]
      properties:
        soldDate:
          type: string
          format: date
          description: Date the listing was observed or reported sold
        amount:
          type: ["number", "null"]
          description: Sale / last ask amount in minor-unscaled currency units (KRW)
        currency:
          type: ["string", "null"]
        registered:
          type: ["string", "null"]
          description: First registration date (YYYY-MM-DD)
        provider:
          type: ["string", "null"]
        sourceListingId:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
    AccidentRow:
      type: object
      required: [date, type]
      properties:
        date:
          type: string
          format: date
          description: Accident or damage date (YYYY-MM-DD)
        type:
          type: string
          enum: [accident, flood_damage, damage]
          description: Normalized accident category
        category:
          type: ["string", "null"]
          description: primary | secondary | flood | insurance type label when known
        damage:
          type: ["string", "null"]
          description: Damage area or label (e.g. Side, Front End)
        description:
          type: ["string", "null"]
        repairTotal:
          type: ["number", "null"]
          description: Total repair cost when reported (usually KRW)
        insuranceBenefit:
          type: ["number", "null"]
          description: Insurance payout when reported
        currency:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        mileageKm:
          type: ["number", "null"]
          description: Odometer in km when recorded on this accident row
        mileageMiles:
          type: ["number", "null"]
    SalvageRecord:
      type: object
      required: [salvage]
      description: >
        US/Canada title brand summary. Present only when a vehicle/auction title
        (or equivalent status) was observed. Korean market cars omit this field.
      properties:
        salvage:
          type: boolean
          description: true when title/status indicates salvage (or junk/rebuilt/COD); false for clean/clear titles
        title:
          type: ["string", "null"]
          description: Primary title text (e.g. Tx - Salvage Vehicle Title)
        detailedTitle:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
          description: Title state code when parsed (TX, CT, …)
        status:
          type: ["string", "null"]
        date:
          type: ["string", "null"]
          format: date
        source:
          type: ["string", "null"]
    VehicleExtraRow:
      type: object
      required: [key, label, value]
      description: >
        Static auction/listing specs (keys, condition, airbags, …) — not timeline events.
        Omitted when no extra fields exist for the VIN.
      properties:
        key:
          type: string
          description: Machine-readable field id (keys, key_status, condition, …)
        label:
          type: string
          description: Human-readable label
        value:
          type: string
        source:
          type: ["string", "null"]
          description: Provider or origin (copart, iaa, salvagebid, …)
        observedAt:
          type: ["string", "null"]
          format: date
...
    MileageHistoryRow:
      type: object
      required: [date, mileageKm, mileageMiles, latest]
      description: >
        Deduplicated odometer reading collected from listings, owners, accidents,
        inspections, or any other VIN tab that recorded mileage plus a date.
        The most recent dated reading is tagged latest.
      properties:
        date:
          type: string
          format: date
        mileageKm:
          type: number
        mileageMiles:
          type: number
        kind:
          type: string
          enum: [listing, owner, accident, inspection, sale, salvage, other]
        source:
          type: ["string", "null"]
        sources:
          type: array
          items:
            type: string
        latest:
          type: boolean
        tag:
          type: string
          enum: [latest]
          description: Set to latest on the most recent dated reading

    # ── Listings ─────────────────────────────────────────────────────────────
...
    PublicErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: INVALID_VIN
            message:
              type: string
              description: Human-readable description
...
    VinCheckEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          required: [vin, exists, providers, hasHistory]
          properties:
            vin:
              type: string
              description: Normalized (uppercase) VIN
              example: KNAGM4A72E5158989
            exists:
              type: boolean
              description: Whether this VIN is in the database
            providers:
              type: array
              items:
                type: string
              description: Internal names of providers that have data for this VIN
              example: [encar]
            hasHistory:
              type: boolean
              description: Whether price/mileage observation history exists
    VinHistoryEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          required: [vin, vehicle, sources, listings, observations, events, photos]
          properties:
            vin:
              type: string
              example: KNAGM4A72E5158989
            vehicle:
              $ref: "#/components/schemas/VinVehicle"
            sources:
              type: array
              items:
                $ref: "#/components/schemas/VinSource"
            listings:
              type: array
              items:
                $ref: "#/components/schemas/VinListing"
            observations:
              type: array
              items:
                $ref: "#/components/schemas/VinObservation"
            events:
              type: array
              items:
                $ref: "#/components/schemas/VinEvent"
            auctionSales:
              type: array
              items:
                $ref: "#/components/schemas/AuctionSaleRow"
            accidents:
              type: array
              items:
                $ref: "#/components/schemas/AccidentRow"
            salvage:
              oneOf:
                - $ref: "#/components/schemas/SalvageRecord"
                - type: "null"
            extra:
              type: array
              items:
                $ref: "#/components/schemas/VehicleExtraRow"
              description: US auction lot specs (keys, condition, …). Omitted when empty.
            mileageHistory:
              type: array
              items:
                $ref: "#/components/schemas/MileageHistoryRow"
            photos:
              type: array
              items:
                $ref: "#/components/schemas/VinPhoto"
        meta:
          type: object
          required: [durationMs]
          properties:
            durationMs:
              type: number
              description: Server-side processing time in milliseconds
            creditCharged:
              type: number
              description: Credits charged for this successful retrieve (1 at current credit price; 0 for test VINs)
    VinVehicle:
      type: object
      properties:
        make:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        year:
          type: ["number", "null"]
        trim:
          type: ["string", "null"]
        bodyType:
          type: ["string", "null"]
        fuelType:
          type: ["string", "null"]
        transmission:
          type: ["string", "null"]
        driveType:
          type: ["string", "null"]
        engineDisplacement:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        currentKnownMileage:
          type: ["number", "null"]
        lastSeenAt:
          type: ["string", "null"]
          format: date-time
    VinSource:
      type: object
      required: [providerId, internalName, name]
      properties:
        providerId:
          type: number
        internalName:
          type: string
          description: Machine-readable provider identifier
          example: encar
        name:
          type: string
          description: Human-readable provider name
          example: Encar
    VinListing:
      type: object
      required: [id, providerId, sourceId, isActive, firstSeenAt, lastSeenAt]
      properties:
        id:
          type: number
        providerId:
          type: number
        sourceId:
          type: string
          description: Provider's own listing identifier
        sourceUrl:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        priceAmount:
          type: ["number", "null"]
          description: Price in minor currency units (e.g. cents for USD, won for KRW)
        priceCurrency:
          type: ["string", "null"]
          example: KRW
        mileage:
          type: ["number", "null"]
        mileageUnit:
          type: ["string", "null"]
          example: km
        location:
          type: ["string", "null"]
        isActive:
          type: boolean
        firstSeenAt:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time
    VinObservation:
      type: object
      required: [id, providerId, observedAt]
      properties:
        id:
          type: number
        providerId:
          type: number
        priceAmount:
          type: ["number", "null"]
        priceCurrency:
          type: ["string", "null"]
        mileage:
          type: ["number", "null"]
        mileageUnit:
          type: ["string", "null"]
        listingStatus:
          type: ["string", "null"]
          enum: [active, sold, expired, null]
        location:
          type: ["string", "null"]
        observedAt:
          type: string
          format: date-time
    VinEvent:
      type: object
      required: [id, eventType, occurredAt]
      properties:
        id:
          type: number
        eventType:
          type: string
          enum: [price_change, status_change, new_listing, delisted]
        description:
          type: ["string", "null"]
        occurredAt:
          type: string
          format: date-time
    VinPhoto:
      type: object
      required: [id, sourceUrl, isPrimary, sortOrder]
      properties:
        id:
          type: number
        sourceUrl:
          type: string
        storedPath:
          type: ["string", "null"]
        isPrimary:
          type: boolean
        sortOrder:
          type: number

    # ── Settings ──────────────────────────────────────────────────────────────
...